feat(rules): switch to API - part 2
